:recycle: (PC-6706) refacto Bloc reusable for Duo choice
